Features
1. Make variables set and defined in one keyword instead of two

2. Split interpreter into multiple files
