+++ cfgl.config
@@ -6,8 +6,8 @@ PRINT = PRINT
 FUNC_DEF = FUNC
 FUNC_CALL = CALL
 FUNC_END = ENDFUNC
-VAR_DEF = VAR
-VAR_ASSIGN = SET
+AS = AS
+VAR = VAR
 ARITHMETIC = CALC
 STR = str
 IMPORT_LIB = USE
