+++ cleanup.py
@@ -1,5 +1,6 @@
 import ast
 import black
+import os
 
 def remove_comments_and_format(file_path):
     with open(file_path, "r") as source:
@@ -12,4 +13,9 @@ def remove_comments_and_format(file_path):
     with open(file_path, "w") as source:
         source.write(formatted_code)
 
-remove_comments_and_format("main.py")
+# Get all Python files in the 'configlang/' directory
+python_files = [f for f in os.listdir('configlang/') if f.endswith('.py')]
+
+# Call remove_comments_and_format on each Python file
+for file in python_files:
+    remove_comments_and_format(f"configlang/{file}")
